+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/ConnectivityStateChangeCause.java
@@ -1,0 +1,6 @@
+package com.raoulvdberge.refinedstorage.apiimpl.network.node;
+
+public enum ConnectivityStateChangeCause {
+    GRAPH_CHANGE,
+    REDSTONE_MODE_OR_NETWORK_ENERGY_CHANGE
+}