+++ src/main/java/com/raoulvdberge/refinedstorage/tile/grid/portable/PortableGrid.java
@@ -11,6 +11,7 @@ import com.raoulvdberge.refinedstorage.api.storage.AccessType;
 import com.raoulvdberge.refinedstorage.api.storage.StorageType;
 import com.raoulvdberge.refinedstorage.api.storage.cache.IStorageCache;
 import com.raoulvdberge.refinedstorage.api.storage.cache.IStorageCacheListener;
+import com.raoulvdberge.refinedstorage.api.storage.cache.InvalidateCause;
 import com.raoulvdberge.refinedstorage.api.storage.disk.IStorageDisk;
 import com.raoulvdberge.refinedstorage.api.storage.disk.IStorageDiskContainerContext;
 import com.raoulvdberge.refinedstorage.api.storage.disk.IStorageDiskProvider;
@@ -129,7 +130,7 @@ public class PortableGrid implements IGrid, IPortableGrid, IStorageDiskContainer
                 }
 
                 if (cache != null) {
-                    cache.invalidate();
+                    cache.invalidate(InvalidateCause.DISK_INVENTORY_CHANGED);
                 }
 
                 StackUtils.writeItems(handler, 4, stack.getTag());